backend/requirements/base.txt
Django>=4.2,<5
psycopg2-binary
celery
redis
coverage
djangorestframework

